Move Lua helper functions to separate module
